Easing requirements on composer.json and requiring stable, adding symfony/intl component; by default not requiring SSL certification for requests
The StatsProd endpoints are failing SSL certification (from the NBA side); since the endpoints for this library all point to NBA domains, it should
be safe to not check the certificates, so by default do not check this. Also, if the local php intl component is not installed there will be Locale
errors on the unit tests, so require the symfony/intl component to resolve this.